Build on clang, which the tests did not.
Three braced initialisers narrowed an int root into KeySig's uint8_t. GCC
warns; clang makes it an error, so the macOS job had been red since the
tier rewrite landed and I had been reporting CI green from a stale look.

Also silences a nodiscard warning on a parseScl call whose return really is
not wanted -- that case is about the name line and the scale is checked
elsewhere.

Verified by building with clang locally rather than pushing and watching:
149 cases, 116199 assertions, no warnings.

+++ test/NoteStrengthTests.cpp
@@ -32,7 +32,7 @@ TEST_CASE("with no chord, the order over scale tones is unchanged",
           "[strength][contract]") {
   for (int root = 0; root < 12; ++root)
     for (int brightness = kLocrian; brightness <= kLydian; ++brightness) {
-      const KeySig k{root, static_cast<int8_t>(brightness), {},
+      const KeySig k{static_cast<uint8_t>(root), static_cast<int8_t>(brightness), {},
                      ScaleType::Diatonic};
       const uint16_t mask = pcMask(k);
       for (int a = 0; a < 12; ++a)
@@ -295,7 +295,7 @@ TEST_CASE("strength stays finite and ordered for every input",
   const auto chord = cMaj7();
   for (int root = 0; root < 12; ++root)
     for (int pc = -24; pc <= 36; ++pc) {
-      const KeySig k{root, kDorian, {}, ScaleType::Diatonic};
+      const KeySig k{static_cast<uint8_t>(root), kDorian, {}, ScaleType::Diatonic};
       const int r = noteStrength(k, pc, chord);
       INFO("root " << root << " pc " << pc << " rank " << r);
       REQUIRE(r >= 0);
@@ -360,7 +360,7 @@ TEST_CASE("the key root sorts first among key-centred notes anyway",
           "[strength][order]") {
   for (int root = 0; root < 12; ++root)
     for (int brightness = kLocrian; brightness <= kLydian; ++brightness) {
-      const KeySig k{root, static_cast<int8_t>(brightness), {},
+      const KeySig k{static_cast<uint8_t>(root), static_cast<int8_t>(brightness), {},
                      ScaleType::Diatonic};
       const int rootRank = noteStrengthRank(k, root);
       REQUIRE(rootRank == 0);
